New tasks commit

+++ src/main/java/nair_khachatryan/shape/Triangle.java
@@ -2,8 +2,8 @@ package nair_khachatryan.shape;
 
 public class Triangle extends Shape {
 
-    public float calculateShapeArea(float base, float hight) {
-        float result = (base * hight) / 2;
+    public double calculateShapeArea(double base, double hight) {
+        double result = (base * hight) / 2;
         System.out.println("The area of the triangle is equal to " + result);
         return result;
     }
